+++ compiler/ml/typedtree.mli
@@ -507,6 +507,7 @@ and type_kind =
 and label_declaration = {
   ld_id: Ident.t;
   ld_name: string loc;
+  ld_runtime_name: string option;
   ld_mutable: mutable_flag;
   ld_optional: bool;
   ld_type: core_type;